naviguide-api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
httpx==0.27.0
searoute==1.4.3
geojson==3.1.0
pandas==2.3.3
copernicusmarine==2.2.4
geographiclib==2.0
python-dotenv==1.0.0
numpy>=1.24.0
xarray>=2023.0.0
shapely>=2.0.0
global-land-mask>=0.5.0
# Simulation Mode — AI Agents (LangGraph + Anthropic Claude)
langgraph>=0.2.0
langchain-core>=0.3.0
requests>=2.31.0
anthropic>=0.25.0
google-generativeai>=0.8.0
google-cloud-secret-manager>=2.20.0
Pillow>=10.0.0
